Updated borders in comments CTA box
DES-189

Our use of borders makes it very difficult to style a theme that looks good with all states. Theme developers need a border to separate comments from the content, but the current comments output includes borders. It's possible to make it look ok in most of the time, but it's difficult to work with. This change updates borders from the comments CTA box with the following rules:

- Logged out without comments: no borders
- Logged out with comments: top border only
- Logged in without comments: no borders
- Logged in with comments: no borders